Island: Move schema parsing to repository utils
Issue: #2783
