+++ monkey/monkey_island/cc/repositories/utils/__init__.py
@@ -1,2 +1,3 @@
 from .mongo_encoder import mongo_dot_decoder, mongo_dot_encoder
 from .machine_repository_init import initialize_machine_repository
+from .agent_configuration_schema_parser import AgentConfigurationSchemaParser
